separate main app file
